requirements.txt
flet
pyqrcode
pypng
